enable upload/download db backup to azure

+++ infrastructure/deployment/scripts/DBRestore.sh
@@ -1,13 +1,26 @@
 #!/bin/bash
 
-# Refer to file name passed in through Azure
-file=$1
+# Clean backups folder
+rm -rf /home/umigrate/backups/*.tar
 
-# Delete the current DB
-sudo -u postgres psql -c "DROP DATABASE umigratedb;"
+# Intake file name
+fileName=""
+if [ -z "$1" ]
+	then
+		echo NO FILE NAME SPECIFIED, EXITING...
+		exit -1
+	else
+  		fileName=$1
 
-# Create a new DB
-sudo -u postgres psql -c "CREATE DATABASE umigratedb;"
+		#Download the file from Azure
+		/home/umigrate/maintenance/azcopy/azcopy copy --recursive "https://umigratefilestorage.blob.core.windows.net/databasebackups/${fileName}${SAS_TOKEN}" /home/umigrate/backups/
 
-# Restoring the DB
-sudo -u postgres pg_restore -d umigratedb $file
+		#Delete the current DB
+		sudo -u postgres psql -c "DROP DATABASE umigratedb;"
+
+  		#Create a new DB
+		sudo -u postgres psql -c "CREATE DATABASE umigratedb;"
+
+  		#Restoring the DB
+  		sudo -u postgres pg_restore -d umigratedb /home/umigrate/backups/$fileName
+fi
